+++ src/HttpFactory/DiactorosHttpFactory.php
@@ -2,101 +2,19 @@
 
 namespace Simply\Application\HttpFactory;
 
-use Psr\Http\Message\RequestInterface;
-use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
-use Psr\Http\Message\StreamInterface;
-use Psr\Http\Message\UploadedFileInterface;
-use Psr\Http\Message\UriInterface;
-use Zend\Diactoros\RequestFactory;
-use Zend\Diactoros\ResponseFactory;
 use Zend\Diactoros\ServerRequestFactory;
-use Zend\Diactoros\StreamFactory;
-use Zend\Diactoros\UploadedFileFactory;
-use Zend\Diactoros\UriFactory;
 
 /**
- * A http factory that uses Zend's Diactoros library to implement the requests and responses.
+ * A http factory that uses Zend's Diactoros library to generate the server request.
  * @author Riikka Kalliomäki <riikka.kalliomaki@gmail.com>
  * @copyright Copyright (c) 2018 Riikka Kalliomäki
  * @license http://opensource.org/licenses/mit-license.php MIT License
  */
 class DiactorosHttpFactory implements HttpFactoryInterface
 {
-    private $requestFactory;
-    private $responseFactory;
-    private $serverRequestFactory;
-    private $streamFactory;
-    private $uploadedFileFactory;
-    private $uriFactory;
-
-    public function __construct()
-    {
-        $this->requestFactory = new RequestFactory();
-        $this->responseFactory = new ResponseFactory();
-        $this->serverRequestFactory = new ServerRequestFactory();
-        $this->streamFactory = new StreamFactory();
-        $this->uploadedFileFactory = new UploadedFileFactory();
-        $this->uriFactory = new UriFactory();
-    }
-
-    /** {@inheritdoc} */
-    public function createRequest(string $method, $uri): RequestInterface
-    {
-        return $this->requestFactory->createRequest($method, $uri);
-    }
-
-    /** {@inheritdoc} */
-    public function createResponse(int $code = 200, string $reasonPhrase = ''): ResponseInterface
-    {
-        return $this->responseFactory->createResponse($code, $reasonPhrase);
-    }
-
-    /** {@inheritdoc} */
-    public function createServerRequest(string $method, $uri, array $serverParams = []): ServerRequestInterface
-    {
-        return $this->serverRequestFactory->createServerRequest($method, $uri, $serverParams);
-    }
-
-    /** {@inheritdoc} */
     public function createServerRequestFromGlobals(): ServerRequestInterface
     {
         return ServerRequestFactory::fromGlobals();
-    }
-
-    /** {@inheritdoc} */
-    public function createStream(string $content = ''): StreamInterface
-    {
-        return $this->streamFactory->createStream($content);
-    }
-
-    /** {@inheritdoc} */
-    public function createStreamFromFile(string $filename, string $mode = 'r'): StreamInterface
-    {
-        return $this->streamFactory->createStreamFromFile($filename, $mode);
-    }
-
-    /** {@inheritdoc} */
-    public function createStreamFromResource($resource): StreamInterface
-    {
-        return $this->streamFactory->createStreamFromResource($resource);
-    }
-
-    /** {@inheritdoc} */
-    public function createUploadedFile(
-        StreamInterface $stream,
-        int $size = null,
-        int $error = \UPLOAD_ERR_OK,
-        string $clientFilename = null,
-        string $clientMediaType = null
-    ): UploadedFileInterface {
-        return $this->uploadedFileFactory
-            ->createUploadedFile($stream, $size, $error, $clientFilename, $clientMediaType);
-    }
-
-    /** {@inheritdoc} */
-    public function createUri(string $uri = ''): UriInterface
-    {
-        return $this->uriFactory->createUri($uri);
     }
 }